refactor: topicのtabsをリファクタリング #32
